server/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
stripe==8.4.0
python-dotenv==1.0.1
bcrypt==4.1.2
requests==2.31.0
mysqlclient==2.2.7
PyMySQL==1.1.0
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==2.3.7 